+++ src/wdio/services/TerraService.js
@@ -2,6 +2,7 @@ import chai from 'chai';
 import chaiMethods from './TerraCommands/chai-methods';
 import accessiblity from './TerraCommands/accessiblity';
 import visualRegression from './TerraCommands/visual-regression';
+import validateElement from './TerraCommands/validate-element';
 import Logger from '../../../scripts/utils/logger';
 import SERVICE_DEFAULTS from '../../../config/wdio/services.default-config';
 
@@ -54,6 +55,7 @@ export default class TerraService {
         matchScreenshot: visualRegression.matchScreenshotWithinTolerance,
         themeEachCustomProperty: visualRegression.themeEachCustomProperty,
         themeCombinationOfCustomProperties: visualRegression.themeCombinationOfCustomProperties,
+        validateElement,
       },
     };
     chai.Assertion.addMethod('accessible', chaiMethods.accessible);